+++ CHANGELOG.rst
@@ -7,6 +7,12 @@
 * Fix test suite for docutils 0.22
   (fixed by mitya57, see pull request #377).
 
+* When there are duplicate citations,
+  resolve locally first
+  (reported by drbenvincent, see issue #385).
+
+* Dropped Python 3.9 support (EOL), added Python 3.14 support.
+
 2.6.5 (27 June 2025)
 --------------------